Module cookie::jar

A cookie jar implementation for storing a set of cookies.

This CookieJar type can be used to manage a session of cookies by keeping track of cookies that are added and deleted over time. It provides a method, delta, which will calculate the number of Set-Cookie headers that need to be sent back to a client which tracks the changes in the lifetime of the jar itself.

A cookie jar can also be borrowed to a child cookie jar with new functionality such as automatically signing cookies, storing permanent cookies, etc. This functionality can also be chained together.

Modules

secure

Structs

Child

CookieJar
A jar of cookies for managing a session

Iter
Iterator over the cookies in a cookie jar

Root

Enums

Flavor

Functions

prepare_key

Type Definitions

Read

SigningKey

Write