Struct cookie::EncodedCookie

pub struct EncodedCookie<'a, 'c: 'a>(_);

Wrapper around Cookie whose Display implementation percent-encodes the cookie's name and value.

A value of this type can be obtained via the encoded method on Cookie. This type should only be used for its Display implementation.

This type is only available when the percent-encode feature is enabled.

Example

use cookie::Cookie;

let mut c = Cookie::new("my name", "this; value?");
assert_eq!(&c.encoded().to_string(), "my%20name=this%3B%20value%3F");