Struct cookie::Key

pub struct Key { /* fields omitted */ }

A cryptographic master key for use with Signed and/or Private jars.

This structure encapsulates secure, cryptographic keys for use with both PrivateJar and SignedJar. It can be derived from a single master key via from_master or generated from a secure random source via generate. A single instance of Key can be used for both a PrivateJar and a SignedJar.

This type is only available when the secure feature is enabled.

Methods

impl Key

fn from_master(key: &[u8]) -> Key

Derives new signing/encryption keys from a master key.

The master key must be at least 256-bits (32 bytes). For security, the master key must be cryptographically random. The keys are derived deterministically from the master key.

Panics

Panics if key is less than 32 bytes in length.

Example

use cookie::Key;

let master_key = { /* a cryptographically random key >= 32 bytes */ };

let key = Key::from_master(master_key);

fn generate() -> Key

Generates signing/encryption keys from a secure, random source. Keys are generated nondeterministically.

Panics

Panics if randomness cannot be retrieved from the operating system. See try_generate for a non-panicking version.

Example

use cookie::Key;

let key = Key::generate();

fn try_generate() -> Option<Key>

Attempts to generate signing/encryption keys from a secure, random source. Keys are generated nondeterministically. If randomness cannot be retrieved from the underlying operating system, returns None.

Example

use cookie::Key;

let key = Key::try_generate();

fn signing(&self) -> &[u8]

Returns the raw bytes of a key suitable for signing cookies.

Example

use cookie::Key;

let key = Key::generate();
let signing_key = key.signing();

fn encryption(&self) -> &[u8]

Returns the raw bytes of a key suitable for encrypting cookies.

Example

use cookie::Key;

let key = Key::generate();
let encryption_key = key.encryption();