Struct cookie::delta::DeltaCookie

pub struct DeltaCookie {
    pub cookie: Cookie<'static>,
    pub removed: bool,
}

A DeltaCookie is a helper structure used in a cookie jar. It wraps a Cookie so that it can be hashed and compared purely by name. It further records whether the wrapped cookie is a “removal” cookie, that is, a cookie that when sent to the client removes the named cookie on the client’s machine.

Fields
cookie: Cookie<'static>removed: bool
Implementations

impl DeltaCookie

pub fn added(cookie: Cookie<'static>) -> DeltaCookie

Create a new DeltaCookie that is being added to a jar.

pub fn removed(cookie: Cookie<'static>) -> DeltaCookie

Create a new DeltaCookie that is being removed from a jar. The cookie should be a “removal” cookie.