Struct cookie::CookieBuilder

pub struct CookieBuilder { /* fields omitted */ }

Structure that follows the builder pattern for building Cookie structs.

To construct a cookie:

Call Cookie::build to start building.

Use any of the builder methods to set fields in the cookie.

Call finish to retrieve the built cookie.

Example

extern crate time;

use cookie::Cookie;
use time::Duration;

let cookie: Cookie = Cookie::build("name", "value")
    .domain("www.rust-lang.org")
    .path("/")
    .secure(true)
    .http_only(true)
    .max_age(Duration::days(1))
    .finish();

Methods

impl CookieBuilder

pub fn new<N, V>(name: N, value: V) -> CookieBuilder where
N: Into<Cow<'static, str>>,
V: Into<Cow<'static, str>>,

Creates a new CookieBuilder instance from the given name and value.

This method is typically called indirectly via Cookie::build.

Example

use cookie::Cookie;

let c = Cookie::build("foo", "bar").finish();
assert_eq!(c.name_value(), ("foo", "bar"));

pub fn expires(self, when: Tm) -> CookieBuilder

Sets the expires field in the cookie being built.

Example

extern crate time;

use cookie::Cookie;

let c = Cookie::build("foo", "bar")
    .expires(time::now())
    .finish();

assert!(c.expires().is_some());

pub fn max_age(self, value: Duration) -> CookieBuilder

Sets the max_age field in the cookie being built.

Example

extern crate time;
use time::Duration;

use cookie::Cookie;

let c = Cookie::build("foo", "bar")
    .max_age(Duration::minutes(30))
    .finish();

assert_eq!(c.max_age(), Some(Duration::seconds(30 * 60)));

pub fn domain<D: Into<Cow<'static, str>>>(self, value: D) -> CookieBuilder

Sets the domain field in the cookie being built.

Example

use cookie::Cookie;

let c = Cookie::build("foo", "bar")
    .domain("www.rust-lang.org")
    .finish();

assert_eq!(c.domain(), Some("www.rust-lang.org"));

pub fn path<P: Into<Cow<'static, str>>>(self, path: P) -> CookieBuilder

Sets the path field in the cookie being built.

Example

use cookie::Cookie;

let c = Cookie::build("foo", "bar")
    .path("/")
    .finish();

assert_eq!(c.path(), Some("/"));

pub fn secure(self, value: bool) -> CookieBuilder

Sets the secure field in the cookie being built.

Example

use cookie::Cookie;

let c = Cookie::build("foo", "bar")
    .secure(true)
    .finish();

assert_eq!(c.secure(), Some(true));

pub fn http_only(self, value: bool) -> CookieBuilder

Sets the http_only field in the cookie being built.

Example

use cookie::Cookie;

let c = Cookie::build("foo", "bar")
    .http_only(true)
    .finish();

assert_eq!(c.http_only(), Some(true));

pub fn same_site(self, value: SameSite) -> CookieBuilder

Sets the same_site field in the cookie being built.

Example

use cookie::{Cookie, SameSite};

let c = Cookie::build("foo", "bar")
    .same_site(SameSite::Strict)
    .finish();

assert_eq!(c.same_site(), Some(SameSite::Strict));

pub fn permanent(self) -> CookieBuilder

Makes the cookie being built 'permanent' by extending its expiration and max age 20 years into the future.

Example

extern crate time;

use cookie::Cookie;
use time::Duration;

let c = Cookie::build("foo", "bar")
    .permanent()
    .finish();

assert_eq!(c.max_age(), Some(Duration::days(365 * 20)));

pub fn finish(self) -> Cookie<'static>

Finishes building and returns the built Cookie.

Example

use cookie::Cookie;

let c = Cookie::build("foo", "bar")
    .domain("crates.io")
    .path("/")
    .finish();

assert_eq!(c.name_value(), ("foo", "bar"));
assert_eq!(c.domain(), Some("crates.io"));
assert_eq!(c.path(), Some("/"));